fw/boards/nrf5: use uint16_t for battery capacity hours
Old Pebbles never offered more than 255 hours I guess!

+++ src/fw/board/board_nrf5.h
@@ -229,7 +229,7 @@ typedef struct {
   const uint8_t low_power_threshold;
 
   //! Approximate hours of battery life
-  const uint8_t battery_capacity_hours;
+  const uint16_t battery_capacity_hours;
 } BoardConfigPower;
 
 typedef struct {
